% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_countries.R
\name{document_countries}
\alias{document_countries}
\title{Look up the countries that the journey went through}
\usage{
document_countries(guid_hash, tidy = TRUE)
}
\arguments{
\item{guid_hash}{Journey global unique identifier to look up in the elasticsearch (see \code{document_journey}).}

\item{tidy}{Tidy the results into a nested tibble \link{DEFAULT: TRUE}.}
}
\value{
tibble.
}
\description{
Look up the countries that the journey went through
}
\examples{
# Lookup the for journey.
\dontrun{
document_countries(guid_hash ="704016213cd70d8cdc58")
}

}
