% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_journey.R
\name{document_journey}
\alias{document_journey}
\title{Look up the journey id for a give document}
\usage{
document_journey(guid_hash, tidy = TRUE)
}
\arguments{
\item{guid_hash}{Journey global unique identifier to look up in the elasticsearch.}

\item{tidy}{Tidy the results into a nested tibble \link{DEFAULT: TRUE}.}
}
\value{
tibble.
}
\description{
Look up the journey id for a give document
}
\examples{
# Lookup the for journey.
\dontrun{
document_journey(guid_hash ="704016213cd70d8cdc58")
}

}
