% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_traveller.R
\name{document_traveller}
\alias{document_traveller}
\title{Return the traveller information for given journey id}
\usage{
document_traveller(journey_id, tidy = TRUE)
}
\arguments{
\item{journey_id}{Journey global unique identifier to look up in the elasticsearch (see \code{document_journey}).}

\item{tidy}{Tidy the results into a nested tibble \link{DEFAULT: TRUE}.}
}
\value{
tibble.
}
\description{
Return the traveller information for given journey id
}
\examples{
# Search for traveller.
\dontrun{
document_traveller(journey_id ="adams_j_1823")
}

}
