% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetravelData.R
\name{get_api_key}
\alias{get_api_key}
\title{Retrieve API key}
\usage{
get_api_key()
}
\value{
API key.
}
\description{
Retrieve API key
}
\examples{
\dontrun{
set_api_key("412dd2813fbf9037f5266557bbf5d1f5")
get_api_key()
}
}
