% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_placenames.R
\name{historical_placenames}
\alias{historical_placenames}
\title{Look up historical placenames}
\usage{
historical_placenames(country_iso, tidy = TRUE)
}
\arguments{
\item{country_iso}{Historical place iso.}

\item{tidy}{Tidy the results into a nested tibble \link{DEFAULT: TRUE}.}
}
\value{
tibble.
}
\description{
Look up historical placenames
}
\examples{
# Search for historical placename
\dontrun{
historical_placenames(country_iso ="KEN")
}

}
