% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parser_meta}
\alias{parser_meta}
\title{Helps to parse the output from the query API}
\usage{
parser_meta(x, full_text = FALSE)
}
\arguments{
\item{x}{A response object.}
}
\value{
tibble
}
\description{
Helps to parse the output from the query API
}
