% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_hits.R
\name{query_hits}
\alias{query_hits}
\title{Search the database and return how many hits}
\usage{
query_hits(query)
}
\arguments{
\item{query}{Specify the text query to send to elasticsearch.}
}
\value{
tibble.
}
\description{
Search the database and return how many hits
}
\examples{
# How many hits were there for a term.
\dontrun{
query_hits(
   query ="cotton" 
   )
}

}
