% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_text.R
\name{query_text}
\alias{query_text}
\title{Search the database}
\usage{
query_text(query, from_pos = 0, tidy = TRUE)
}
\arguments{
\item{query}{Specify the text query to send to elasticsearch.}

\item{from_pos}{\link{OPTIONAL} Parameter to paginate through result.}

\item{tidy}{Tidy the results into a nested tibble \link{DEFAULT: TRUE}.}
}
\value{
tibble.
}
\description{
Search the database
}
\examples{
# Look for the term cotton.
\dontrun{
query_text(
   query ="cotton", 
   from_pos = 0, 
   tidy = TRUE
   )
}

}
