% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetravelData.R
\name{set_api_key}
\alias{set_api_key}
\title{Set API key}
\usage{
set_api_key(api_key)
}
\arguments{
\item{api_key}{API key.}
}
\description{
Specify the key for accessing the timetravelData API.
Automatically determines which type of key is being used and authenticates appropriately.
}
\examples{
# Authenticate using an API key.
set_api_key("412dd2813fbf9037f5266557bbf5d1f5")

}
