% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetravelData.R
\name{set_base_url}
\alias{set_base_url}
\title{Set base URL for API  when testing}
\usage{
set_base_url(url)
}
\arguments{
\item{url}{URL to access API.}
}
\description{
Set base URL for API  when testing
}
\examples{
# Use local server
set_base_url("http://0.0.0.0:8080/")
}
