% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetravelData.R
\name{set_server}
\alias{set_server}
\title{Set IP and port for API server when testing}
\usage{
set_server(ip, port)
}
\arguments{
\item{ip}{IP address for API server.}

\item{port}{Port on which API is running.}
}
\description{
Set IP and port for API server when testing
}
\examples{
# Use local server
set_server("0.0.0.0", 8080)
}
